### 2021-06-08 #3345 @whedon
[PRE REVIEW]: pyDeltaRCM: a flexible numerical delta model
**Submitting author:** @amoodie (<a href="http://orcid.org/0000-0002-6745-036X">Andrew Moodie</a>)
**Repository:** <a href="https://github.com/DeltaRCM/pyDeltaRCM" target ="_blank">https://github.com/DeltaRCM/pyDeltaRCM</a>
**Version:** v2.0
