+++ docs/source/_resources/joss/paper.bib
@@ -81,7 +81,7 @@
 	journal = {Journal of Geophysical Research: Earth Surface},
 	volume = {126},
 	number = {4},
-	doi = {https://doi.org/10.1029/2020JF005706},
+	doi = {10.1029/2020JF005706},
 	year = {2021}
 }
 
