### 2021-06-08 #3345 @amoodie
Awesome, thank you @kbarnhart!

I'll get the malformed doi whedon pointed out above fixed, along with anything you point out that needs to be corrected.

Some potential reviewers:
* Katherine Ratliff (river avulsion and floodplain evolution model `rafem` developer) `katmratliff`
* Nathan Lyons (Python model development) `nathanlyons`
* Zoltán Sylvester (Python model development, surface processes) `zsylvester`
