### 2021-06-10 9735b32
increment version for 2.0 release on JOSS review
