### 2021-06-11 #3345 @kbarnhart
@amoodie a few things follow. All are minor so feel free to address them along with reviewer comments:

1. CI failing. Looking at it, it seems like its not failing tests, but just failing coverage upload (https://github.com/DeltaRCM/pyDeltaRCM/runs/2776370520?check_suite_focus=true).
2. Some reference fixes
    * The invalid DOI you noted.
    * Reference ambiguity (M. Liang vs Man Liang, to fix make sure the name is identical in the bib file)
    * In the BMI reference, Eric Hutton's second middle initial should be capitalized.

Thanks for your thoughts on reviewers.
