+++ docs/source/_resources/joss/paper.bib
@@ -1,5 +1,5 @@
 @article{liang_reduced_1_2015,
-	author = {Liang, M. and Voller, V. R. and Paola, C.},
+	author = {Liang, Man and Voller, Vaughan R. and Paola, Chris},
 	title = {A reduced-complexity model for river delta formation -- {P}art 1: Modeling deltas with channel dynamics},
 	journal = {Earth Surface Dynamics},
 	volume = {3},
@@ -10,7 +10,7 @@
 }
 
 @article{liang_reduced_2_2015,
-	author = {Liang, M. and Geleynse, N. and Edmonds, D. A. and Passalacqua, P.},
+	author = {Liang, Man and Geleynse, Nathanael. and Edmonds, Douglas A. and Passalacqua, Paola},
 	title = {A reduced-complexity model for river delta formation -- {P}art 2: Assessment of the flow routing scheme},
 	journal = {Earth Surface Dynamics},
 	volume = {3},
@@ -126,7 +126,7 @@
 	volume = {5},
 	number = {51},
 	pages = {2317},
-	author = {Eric W.h. Hutton and Mark D. Piper and Gregory E. Tucker},
+	author = {Eric {W. H.} Hutton and Mark D. Piper and Gregory E. Tucker},
 	title = {The Basic Model Interface 2.0: A standard interface for coupling numerical models in the geosciences},
 	journal = {Journal of Open Source Software}
 }
