### 2021-06-11 #3345 @amoodie
Alright, should be all good to go now. Just released v2.0.0, which is the version to review. :+1:
