### 2021-06-23 #3398 @kbarnhart
@amoodie could you see about the PDF compilation error? If you can't figure it out, let me know and I can tag one of the whedon experts on the JOSS editorial team to help out.
