### 2021-06-23 #3398 @amoodie
Hi @kbarnhart, I've looked into that error, but have not identified any leads to explain it. The paper compiles [on the preview service](http://res.cloudinary.com/hkvhhndao/image/upload/v1624470890/h6jbfbnsz2rpnkblsgc8.pdf) without issue, and so I do think we need the help of an expert.

Thanks reviewers for your time, and looking forward to your feedback.
