### 2021-06-23 #3398 @kbarnhart
@arfon any recommendations on addressing this PDF compile error (or who else I should ping about it).
