### 2021-06-24 #3398 @arfon
@kbarnhart – well that was a new one for me 😸. It turns out the URL you have in your JOSS bio `https://www.usgs.gov/staff-profiles/katherine-barnhart?qt-staff_profile_science_products=3#qt-staff_profile_science_products` was causing some issues with the paper compilation (as we link to editors in the JOSS paper). For now I've changed it to `https://www.usgs.gov/staff-profiles/katherine-barnhart` which seems to redirect to `https://www.usgs.gov/staff-profiles/katherine-barnhart?qt-staff_profile_science_products=3#qt-staff_profile_science_products` anyway but I'll need to look into fixing this in the LaTeX template.
