### 2021-06-29 #3398 @kbarnhart
@salterg - thanks for agreeing to review this submission to JOSS. In the prior comments there is a checklist that you can use to guide you through your review.

We have an automatic reminder set up to ask you how the review is going. At present we request that reviewers complete their reviews within 6 weeks (August 10, 2021). JOSS is trying to be mindful of changes people have experienced due to COVID-19.

As you work through your review, if there are any issues that come up, please make an issue in the pyDeltaRCM repository, and link to this issue (paste openjournals/joss-reviews/issues/3398 into the issue you create). That way most of the discussion can occur on in-repo issues.

If you have any questions, please let me know (tag me here or email krbarnhart@usgs.gov).
