### 2021-07-10 DeltaRCM/pyDeltaRCM#215 @jhnienhuis
JOSS paper review minor issues
Hello @amoodie ,
First of all, apologies if I'm doing this review incorrectly or if i'm not following github customs. This is certainly a learning experience for me.

I noticed two smalls things when trying to follow your tutorial (https://deltarcm.org/pyDeltaRCM/guides/10min.html):
## 1)

>>> import pyDeltaRCM
>>> default_delta = pyDeltaRCM.DeltaModel()
>>> for _ in range(0, 5):
>>>  delta.update()
>>> delta.finalize()

default_delta is changed to delta, so it will give an error message if you're blindly following these steps.

## 2)

where can I find this file '10min_tutorial.yaml'?

Best wishes,
Jaap
