### 2021-07-10 #3398 @jhnienhuis
@kbarnhart,

thanks again for inviting me to this review. Its been fun exploring github and python!

I noticed two small things:
1) I noticed this link is expired. do I need to accept anything still?
Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations

2)  I tried to find the paper pdf on the JOSS webpage because I could not find it anywhere on github. I eventually found it, but by doing that also noticed the search function doesn't really work properly.
![image](https://user-images.githubusercontent.com/34747577/125174787-16f81d00-e1c8-11eb-9177-93e4c0d4a667.png)
for example, when i search for "@amoodie" or pydeltarcm is doesn't find the pdf of this manuscript: https://joss.theoj.org/papers/search?q=pydeltarcm

Perhaps you can forward this to the appropriate webpage developer?

Best,
Jaap
