### 2021-07-12 a01886c
fix delta name and elaborate on how to create yaml file
