### 2021-07-12 DeltaRCM/pyDeltaRCM#215 @amoodie
Hi Jaap,

You're a github pro as far as I can see! Thanks for pointing out these issues.

I've fixed the first error. For number 2), there is no file. Our intention is to have the user learn how to create the YAML file themselves, as knowing how to set up this file is pretty central to using the model. However, I see that wanting the user to make this file was obviously not clear in the documentation. I've updated the text in that section to hopefully make it clear that the user should make the file, and a brief description for how to make the file.
