### 2021-07-12 #3398 @kbarnhart
@jhnienhuis above is a new link.

A link to the paper is in one of the [above comments](https://github.com/openjournals/joss-reviews/issues/3398#issuecomment-867507889), and we can re-build the current version anytime by writing `@whedon generate pdf` in a comment.

However, you also bring up a great point regarding search capabilities on the website. @arfon, do we expect search results on the website (https://joss.theoj.org/papers) to yield results if the paper is active rather than published?
