### 2021-07-12 #3398 @arfon
> However, you also bring up a great point regarding search capabilities on the website. @arfon, do we expect search results on the website (https://joss.theoj.org/papers) to yield results if the paper is active rather than published?

No, search only lists papers that are published.
