### 2021-07-20 DeltaRCM/pyDeltaRCM#217 @kbarnhart
Salter JOSS Review
@amoodie attached is a word doc with a review from @salterg. Thanks for providing this thoughtful and constructive review.

[salterjossreview.docx](https://github.com/DeltaRCM/pyDeltaRCM/files/6851740/salterjossreview.docx)

openjournals/joss-reviews/issues/3398
