### 2021-07-22 DeltaRCM/pyDeltaRCM#215 @jhnienhuis
Thanks @amoodie!, your answers make sense, although (from a end-user perspective) I usually run the default and adapt files where needed instead of trying to make new files myself. I see where you're coming from though.
Cheers,
Jaap
