### 2021-07-24 025d1b7
update a number of text points and add links to other install instructions.
