+++ _sources/meta/installing.rst.txt
@@ -2,8 +2,8 @@
 Installing
 ************
 
-We recommend installing pyDeltaRCM in a virtual environment.
-That said, *pyDeltaRCM* depends on a small number of packages (:ref:`list of dependencies <dependencies-list>`), many of which are likely already in a Python user/developer's regular use, so it's probably safe to install it in your base environment, too.
+We recommend installing *pyDeltaRCM* in a virtual environment.
+That said, *pyDeltaRCM* depends on a small number of packages (:ref:`list of dependencies <dependencies-list>`), many of which are likely already in a Python user/developer's regular use, so it's probably safe to install *pyDeltaRCM* in your base environment, too.
 
 
 Installing
@@ -13,18 +13,20 @@ We describe installation flavors for both users and developers below.
 
 .. hint::
 
-    If you are looking to make any modifications to the code base, you should follow the developer instructions.
+    If you are looking to make any modifications to the model source code, you should follow the developer instructions.
 
 We suggest using the Anaconda Python distribution, which you can obtain via `the project website <https://www.anaconda.com/products/individual>`_.
 
 Before proceeding, you may wish to create a virtual environment for the *pyDeltaRCM* project.
-With Anaconda:
+With Anaconda on Linux:
 
 .. code:: console
 
     $ conda create -n deltarcm
+    $ conda activate deltarcm
 
-See `this helpful guide <https://packaging.python.org/guides/installing-using-pip-and-virtual-environments/#creating-a-virtual-environment>`_ for creating virtual environments with `venv` if you do not use Anaconda.
+For more informtaion, see `this guide <https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#activating-an-environment>`_ for help on creating and activating a virtual environment with Anaconda on other platforms.
+See `this guide <https://packaging.python.org/guides/installing-using-pip-and-virtual-environments/#creating-a-virtual-environment>`_ for creating virtual environments with `venv` if you do not use Anaconda.
 
 
 User installation
@@ -35,6 +37,10 @@ For a user installation, simply install from the pypi package repository:
 .. code:: console
 
     $ pip install pyDeltaRCM
+
+.. note::
+
+    You may need to `first install <https://pip.pypa.io/en/stable/installing/>`_ `pip`.
 
 
 .. _dev-install:
@@ -49,7 +55,8 @@ This will allow you to submit suggestions and contribute to *pyDeltaRCM*.
 
     You do not *need* to create a fork if your are just testing, but it may save you time and headache down the road. If you choose not to, just use the main repository url below (https://github.com/DeltaRCM/pyDeltaRCM.git).
 
-First download or clone your fork of the project:
+First, you will need to `install git <https://git-scm.com/book/en/v2/Getting-Started-Installing-Git>`_  if you do not already have it.
+Then, download or clone your fork of the project:
 
 .. code:: console
 
