### 2021-08-01 3fd4ebe
imrpove clarity of best practices and expand on the contributing guide
