### 2021-08-04 DeltaRCM/pyDeltaRCM#219 @zsylvester
JOSS Review and related comments
Hi @amoodie and coauthors,

I have enjoyed reading your paper and have enjoyed exploring pyDeltaRCM even more. Clearly a ton of good work went into this; I think this is a great delta model and Python package and I am hoping to use it in the future. I am convinced that many others will find it useful as well. I have not gone through all the documentation, tests, and examples in detail -- there is a lot to explore here and I think that going through every detail is beyond the scope of a review of a short JOSS paper (but correct me @kbarnhart if I am wrong). I have a few comments below but I don't think that addressing any of these are critical for publishing the paper in JOSS; however, hopefully some of them are useful for improving the package. Here is the list:

- Maybe this just reflects my addiction to Jupyter notebooks, but I think it would be useful to provide some example notebooks (or at least one). For a lot of people, it is easier to get started with a package if there is a ready-to-go notebook in the repository.
- The 10-minute tutorial is great, but it would be nice if you made it a bit more obvious how one can generate a nice-looking delta morphology with well-defined channels, e.g., by showing how running the default model for 1000 iterations (which might take a few minutes, but that's OK) gives a nice result.
- The workflow of adding non-default parameters to a minimalistic yaml file works well (although it is not very clear from the documentation that this is a best practice).
- However, I have tried to create an input 'yaml' file with all the parameters and ran into several problems. I have copied the parameter list from [here](https://deltarcm.org/pyDeltaRCM/reference/model/yaml_defaults.html) and found that several of the default values listed there give an error message when I am trying to read the file. Also, those error messages look like "input for Np_water not of the right type in yaml configuration file {self.input_file}. Input type was {type(k).__name__}, but needs to be {expected_type}". I have got it to work in the end, but, long story short, it would be useful to provide an example input file with all the default parameters so that the user can have an overview of all the parameters and start changing them.
- Not clear what should be used in the input parameter file if you don’t want to specify that value (it looks to me like ‘None’ and ‘null’ don’t work) - I guess you just comment out / delete that line?
- In the description of the 'stepmax' parameter ([here](https://deltarcm.org/pyDeltaRCM/reference/model/yaml_defaults.html)), the phrase “stepmax is the maximum number of jumps a parcel can make” is repeated twice.
- In the same section, fix the typos in “oceaninc bounary”.
- The 10-minute tutorial should point out that, if the model is not ‘finalized’ it is possible to update the model with more time steps.
- I understand that dissecting and analyzing the output of pyDeltaRCM in detail is beyond the scope of this package and that you are working on another package to do that (looking forward to trying that package out as well!), but I think it would be helpful (and would make the package more appealing) if you provided a simple example of basic visualizations, e.g., how to save depth grids from the model, read back the netCDF4 file, and display the results. Something as simple as this would save a lot of time for a lot of people: `from netCDF4 import Dataset;
nc = Dataset('pyDeltaRCM_output.nc'); plt.imshow(nc['depth'][100,:,:], cmap='viridis_r')`
- The paper says that the model is compatible with the Basic Model Interface, but the latest version on Github states that it is not.
- The installation instructions: when creating a new conda environment it is a good idea to specify the Python version (at least with my setup).
- It would be good to get a rough idea about performance (e.g., how long does it take to create a 'nice' delta model?). Right now this is not clear at all after reading the paper.
- It might be helpful to mention in the paper the advantages / disadvantages of using a reduced complexity model. Some readers might not realize how hard it is to install and run a more sophisticated morphodynamic model like Delft3D.

That's it for now -- thanks for building this and making it open source!

https://github.com/openjournals/joss-reviews/issues/3398
