### 2021-08-04 DeltaRCM/pyDeltaRCM#219 @kbarnhart
Thanks for this great review @zsylvester !
