### 2021-08-04 #3398 @kbarnhart
@zsylvester @jhnienhuis @salterg thanks for providing your reviews.

@amoodie please ping me here or at krbarnhart@usgs.gov when you think you have addressed the issues raised by the reviewers, or at any point if you need input or guidance from me.
