### 2021-08-06 DeltaRCM/pyDeltaRCM#219 @amoodie
## response to zsylvester review

*I have enjoyed reading your paper and have enjoyed exploring pyDeltaRCM even more. Clearly a ton of good work went into this; I think this is a great delta model and Python package and I am hoping to use it in the future. I am convinced that many others will find it useful as well. I have not gone through all the documentation, tests, and examples in detail -- there is a lot to explore here and I think that going through every detail is beyond the scope of a review of a short JOSS paper (but correct me `kbarnhart` if I am wrong). I have a few comments below but I don't think that addressing any of these are critical for publishing the paper in JOSS; however, hopefully some of them are useful for improving the package. Here is the list:*
Thank you Zoltán for your kind words and for taking the time review our project. We appreciate the suggestions you made, and have responded to each of them below.

- *Maybe this just reflects my addiction to Jupyter notebooks, but I think it would be useful to provide some example notebooks (or at least one). For a lot of people, it is easier to get started with a package if there is a ready-to-go notebook in the repository.*
Thanks for this suggestion. We added a Jupyter Notebook ([link](https://github.com/deltaRCM/pyDeltaRCM/blob/develop/docs/source/examples/simple_example.ipynb)) that includes creating a YAML file, running the model for 1000 timesteps, and loading the output NetCDF file and creating a timeseries plot.

- *The 10-minute tutorial is great, but it would be nice if you made it a bit more obvious how one can generate a nice-looking delta morphology with well-defined channels, e.g., by showing how running the default model for 1000 iterations (which might take a few minutes, but that's OK) gives a nice result.*
Thanks for this suggestion. We debated this idea internally a few months ago, and decided that we wanted to keep our documentation build time less than ~3 minutes. As a result, we can't run the model for more than a few timesteps anywhere in the documentation. We take advantage of the *pyDeltaRCM* [checkpointing feature](https://deltarcm.org/pyDeltaRCM/reference/iteration_tools/index.html?highlight=checkpoint#pyDeltaRCM.iteration_tools.iteration_tools.save_the_checkpoint) to have one pre-computed model result for visualization in the documentation (e.g., [the delta centered on the docs homepage](https://deltarcm.org/pyDeltaRCM/index.html), and this precomputed delta took approximately two hours to run. We prefer not to add a longer run to the built documentation, but we did add a Jupyter Notebook (in response to your suggestion above), which includes a 1000-timestep run ([link](https://github.com/deltaRCM/pyDeltaRCM/blob/develop/docs/source/examples/simple_example.ipynb)).

- *The workflow of adding non-default parameters to a minimalistic yaml file works well (although it is not very clear from the documentation that this is a best practice).*
Thanks for pointing this out. We agree we should emphasize best practices. We have added a callout "important" box to the documentation in the [10-minute Guide here](https://github.com/DeltaRCM/pyDeltaRCM/pull/223/files#diff-94ee9f74808c7bfc24e0f2a9a44d143316f5b3215fbc1689f492494459f184b7R63-R65) and the [User Guide here](https://github.com/DeltaRCM/pyDeltaRCM/pull/223/files#diff-8c24e8fb55c3167b9cfc22723b2e4877f6d937b649e9f1aec52ee349d3d69769R120-R123).

- *However, I have tried to create an input 'yaml' file with all the parameters and ran into several problems. I have copied the parameter list from [here](https://deltarcm.org/pyDeltaRCM/reference/model/yaml_defaults.html) and found that several of the default values listed there give an error message when I am trying to read the file. Also, those error messages look like "`input for Np_water not of the right type in yaml configuration file {self.input_file}. Input type was {type(k).__name__}, but needs to be {expected_type}`". I have got it to work in the end, but, long story short, it would be useful to provide an example input file with all the default parameters so that the user can have an overview of all the parameters and start changing them.*
Thanks for pointing out this issue. We fixed the broken error message [here in #220](https://github.com/DeltaRCM/pyDeltaRCM/pull/220/files#diff-4c514b0f007ec89913d29b4fadd8ba787aece14c69fdb64c19771c02732ca5c3R141-R144). We hope that with the added clarification on best practices (above and below comments) in the documentation, it will be clear how to set parameters in the model.

- *Not clear what should be used in the input parameter file if you don’t want to specify that value (it looks to me like ‘None’ and ‘null’ don’t work) - I guess you just comment out / delete that line?*
Thanks for raising this issue. Yes, you should comment out or remove the line completely; setting `null` does not work. We added a call-out box suggesting to comment out lines to the Advanced Configuration guide ([here in #223](https://github.com/DeltaRCM/pyDeltaRCM/pull/223/files#diff-9d4b646ac605a621bcc34946fb5834cdce4b62ebeaea57135d06b10c1de419bfR93-R96)) and to the User Guide ([here in #223](https://github.com/DeltaRCM/pyDeltaRCM/pull/223/files#diff-8c24e8fb55c3167b9cfc22723b2e4877f6d937b649e9f1aec52ee349d3d69769R120-R123)).

- *In the description of the 'stepmax' parameter ([here](https://deltarcm.org/pyDeltaRCM/reference/model/yaml_defaults.html)), the phrase “stepmax is the maximum number of jumps a parcel can make” is repeated twice.*
Thanks for pointing this out, we revised the text here to make it less repetitive ([here in #223](https://github.com/DeltaRCM/pyDeltaRCM/pull/223/files#diff-a921603a49eabb5796c141c06a9d7d75fdd3471847038fb76dcb90ea46c5c434R1221-R1242))

- *In the same section, fix the typos in “oceaninc bounary”.*
Thanks for catching this typo, we [fixed this in #220](https://github.com/DeltaRCM/pyDeltaRCM/pull/220/files#diff-a921603a49eabb5796c141c06a9d7d75fdd3471847038fb76dcb90ea46c5c434R1227).

- *The 10-minute tutorial should point out that, if the model is not ‘finalized’ it is possible to update the model with more time steps.*
Thanks for this suggestion, we added a note to explain that more updates can be made, until `finalize` is called ([here in #220](https://github.com/DeltaRCM/pyDeltaRCM/pull/220/files#diff-94ee9f74808c7bfc24e0f2a9a44d143316f5b3215fbc1689f492494459f184b7R38-R41)).

- *I understand that dissecting and analyzing the output of pyDeltaRCM in detail is beyond the scope of this package and that you are working on another package to do that (looking forward to trying that package out as well!), but I think it would be helpful (and would make the package more appealing) if you provided a simple example of basic visualizations, e.g., how to save depth grids from the model, read back the netCDF4 file, and display the results. Something as simple as this would save a lot of time for a lot of people: `from netCDF4 import Dataset;
nc = Dataset('pyDeltaRCM_output.nc'); plt.imshow(nc['depth'][100,:,:], cmap='viridis_r')`*
Thanks for this suggestion. We have added a brief explanation for how to connect to the NetCDF file and examine the data. Any analysis is outside the scope of what we want to include in the *pyDeltaRCM* documentation, though. The example is included in the User Guide ([see it rendered here](https://deltarcm.org/pyDeltaRCM/guides/user_guide.html#model-output-file)) and the informational section of the model ([added here in #223](https://github.com/DeltaRCM/pyDeltaRCM/pull/223/files#diff-dd867c1d36f424e0cc309432be6a0c8e0a015319069ac435c82bc63c2e1dc003R63-R83)).

- *The paper says that the model is compatible with the Basic Model Interface, but the latest version on Github states that it is not.*
Thanks for pointing out this confusing language on the Github README. In development, we moved the BMI wrapper outside of the *pyDeltaRCM* repository. So, *pyDeltaRCM* is BMI-compatible [via the wrapper](https://deltarcm.org/BMI_pyDeltaRCM/index.html) and it's [incredibly easy to use](https://deltarcm.org/BMI_pyDeltaRCM/userguide.html). None of us are currently using the BMI wrapper, but we think this JOSS publication would be a good opportunity to advertise to the community that the model is BMI-ready. We updated the README to hopefully make this more clear ([here in #223](https://github.com/DeltaRCM/pyDeltaRCM/pull/223/files#diff-7b3ed02bc73dc06b7db906cf97aa91dec2b2eb21f2d92bc5caa761df5bbc168fR57-R58)).

- *The installation instructions: when creating a new conda environment it is a good idea to specify the Python version (at least with my setup).*
Thanks for this suggestion. We have made this change ([here in #220](https://github.com/DeltaRCM/pyDeltaRCM/pull/220/files#diff-dd47029807d3a686a46a1b8894a143d664450db3150b31fcc08b39d366604965R25)).

- *It would be good to get a rough idea about performance (e.g., how long does it take to create a 'nice' delta model?). Right now this is not clear at all after reading the paper.*
Thanks for this suggestion, we added the run time for the delta in Figure 1 to the caption of that figure ([here in #224](https://github.com/DeltaRCM/pyDeltaRCM/pull/224/files#diff-02c2e7e36d1f0ff426e4a3b39f0ee9aca75ecdde3d7497530222b8d1b0812797R60)) . That simulation took [1 hr 57 minutes](https://github.com/DeltaRCM/pyDeltaRCM/pull/224/files#diff-36d1000034705f225884bf87369502b5b6b485f2f59fb39488e254afbdf14219) on my personal workstation. For the sake of completeness in this answer: computational time depends on parameter choices, and especially the number of grid cells in the model domain. Larger domains with smaller grid cells take exponentially longer to compute, for example more than 12 hours for some cases in a recent research study.

- *It might be helpful to mention in the paper the advantages / disadvantages of using a reduced complexity model. Some readers might not realize how hard it is to install and run a more sophisticated morphodynamic model like Delft3D.*
Thanks for this suggestion. In the paper "Statement of need" section, we say "The DeltaRCM delta model (Liang, Voller, et al., 2015) has gained popularity among geomorphologists due to an attractive balance of computational cost, realism, and interpretability (Larsen et al., 2016)." We believe this gives readers a sense of why the reduced complexity approach is popular, without making the explanation too verbose for the short JOSS paper (e.g., we include the reference to Larsen et al., 2016).

*That's it for now -- thanks for building this and making it open source!*
Thanks again for your thoughtful suggestions, we appreciate the review!
