### 2021-08-06 DeltaRCM/pyDeltaRCM#215 @amoodie
For my own record keeping during revisions:

The issues identified in this review were addressed in #216.
