### 2021-08-06 6c72904
fix error thrown to log during finalize when no output exists.
