### 2021-08-06 8dc9b3a
clarify mass conservation point
