### 2021-08-06 35f8b1c
complete missing doc sections
