### 2021-08-09 f6610f4
add info on slicing
