### 2021-08-09 d8ccf6e
add a jupyter notebook example, update text to explain best practice for yaml:
