### 2021-08-09 b2b4ef0
update getting started page, add links to jupyter, fix broken refs
