+++ docs/source/guides/getting_started.rst
@@ -5,9 +5,21 @@ Getting Started
 ===============
 
 
-This section gives a learn-by-example lesson on how to use pyDeltaRCM. 
+It's easy to get started using *pyDeltaRCM*.
+We recommend starting out with the 10-minute learn-by-example lesson:
 
 .. toctree::
    :maxdepth: 3
 
    10min
+
+
+Return to the main index :ref:`user_documentation` to see all the other documentation available to help you get started.
+
+Some other helpful resources:
+
+* An example `Jupyter Notebook <https://github.com/deltaRCM/pyDeltaRCM/blob/develop/docs/source/examples/simple_example.ipynb>`_ that shows a complete model runs and quick analysis of output file.
+* A comprehensive :doc:`user_guide` on how to use and experiment with the model
+
+
+And when you're ready to start developing, check out the :doc:`developer_guide` and the complete model :doc:`API documentation </reference/index>`.
