### 2021-08-09 7ccdd38
update stepmax docstring
