### 2021-08-09 0cd8095
add an example for how to connect to output file, slice data
