### 2021-08-10 d7db042
update paper caption and figure with new run and log
