### 2021-08-12 #3398 @amoodie
Hi @kbarnhart :wave:

We got great feedback from three reviewers, which mostly revolved around points of clarification within the documentation. We have responded to all of the reviewers' comments and questions in repository issues:

* Zoltán review: https://github.com/DeltaRCM/pyDeltaRCM/issues/219#issuecomment-894295438
* Jaap review: https://github.com/DeltaRCM/pyDeltaRCM/issues/215
* Gerard review: https://github.com/DeltaRCM/pyDeltaRCM/issues/217#issuecomment-885003219

I've also regenerated the article proof above, which includes a minor change based on the reviews. Please let me know when you need anything from me to move forward. Thanks for considering our package in JOSS.
