### 2021-08-18 DeltaRCM/pyDeltaRCM#219 @zsylvester
Thanks @amoodie for the careful responses - all is good as far as I am concerned. I did bump into a question in the meantime (that I could probably find the answer to in the documentation / published papers, but this is probably faster...): if I run the simulation with a smaller grid size (e.g., 10 m instead of 50 m -- I wanted a similar-looking, but higher resolution model), and every other parameter stays the same, I get a very different delta (in terms of channel size, sinuosity, etc.). Is there a best practice to deal with this? Or is there a fundamental dependence of the model on grid cell size?
