### 2021-08-19 DeltaRCM/pyDeltaRCM#219 @amoodie
Hi @zsylvester, thanks again for your review!

Yes, there is a fundamental grid size dependence in the model. @elbeejay has investigated the issue a bit, and he has some evidence that the avulsion dynamics of the system remains approximately the same regardless of the grid size. Still though, we think the cell-size dependence needs to be studied more carefully and don't have any best practices identified. In the original DeltaRCM Liang et al., 2015a,b, 2016a,b papers, the authors used grid sizes of 50 and 100 meters, but didn't comment on the dependence, as far as I can remember. So far, we have chosen according to whatever the coarsest size that resolves the dynamics needed to address the science question (usually 50--150 m).
