### 2021-08-19 #3398 @kbarnhart
@zsylvester thanks for that update and for raising that other issue about grid scale dependince.

@jhnienhuis @salterg -- a friendly reminder to take a moment and revisit your reviews. If the authors have addressed your concerns, please complete your review checklists and indicate that you think the submission is ready to be accepted. If remaining issues exist, please note them in an in-repo issue.
