### 2021-08-25 #3398 @kbarnhart
@amoodie Please do the following
- [ ] Revise the sentence starting at Line 11. Its unclear what "This" refers to (the prior sentence or contents following in its own sentence.
- [ ] Once that change is made, run `@whedon generate pdf` and `@whedon check references`. Presuming both of those complete without error continue to the next steps.
- [ ] Make a tagged release of your software, and report the version tag of the archived version here.

I can then move forward with the next stages of the JOSS process.
