### 2021-09-03 4718a63
add joss badge without line breaks
