### 2021-10-14 1072bf9
make dimensions and variables use the same names, x and y, for the output netcdf file. This is what xarray expects.
