### 2021-10-14 dd7b889
update coordinate specification in model initialization.
