### 2021-10-14 1da8851
implement legacy netcdf method with default as legacy version, note all tests passing here.
