+++ pyDeltaRCM/init_tools.py
@@ -176,6 +176,10 @@ class init_tools(abc.ABC):
         self._input_file_vars = input_file_vars
         self.out_dir = self._input_file_vars['out_dir']
         self.verbose = self._input_file_vars['verbose']
+        if self._input_file_vars['legacy_netcdf']:
+            self._netcdf_coords = ('total_time', 'length', 'width')
+        else:
+            self._netcdf_coords = ('time', 'y', 'x')
 
     def process_input_to_model(self):
         """Process input file to model variables.
@@ -275,6 +279,7 @@ class init_tools(abc.ABC):
                                 self._save_velocity_components)
         if self._save_any_grids:  # always save metadata if saving grids
             self._save_metadata = True
+
         self._is_finalized = False
 
     def set_constants(self):
@@ -599,23 +604,36 @@ class init_tools(abc.ABC):
             self.output_netcdf.source = 'pyDeltaRCM v{ver}'.format(
                 ver=self.__pyDeltaRCM_version__)
 
-            # create master dimensions
-            self.output_netcdf.createDimension('y', self.L)
-            self.output_netcdf.createDimension('x', self.W)
-            self.output_netcdf.createDimension('time', None)
+            # create master dimensions (pulls from `self._netcdf_coords`)
+            self.output_netcdf.createDimension(self._netcdf_coords[1], self.L)
+            self.output_netcdf.createDimension(self._netcdf_coords[2], self.W)
+            self.output_netcdf.createDimension(self._netcdf_coords[0], None)
 
             # create master coordinates (as netCDF variables)
-            x = self.output_netcdf.createVariable(
-                'x', 'f4', ('x'))
-            y = self.output_netcdf.createVariable(
-                'y', 'f4', ('y'))
-            time = self.output_netcdf.createVariable('time', 'f4',
-                                                     ('time',))
-            x.units = 'meters'
-            y.units = 'meters'
+            time = self.output_netcdf.createVariable(
+                'time', 'f4', (self._netcdf_coords[0],))
             time.units = 'second'
-            x[:] = self.xc
-            y[:] = self.yc
+
+            if self._legacy_netcdf:
+                # old format is 2d array x and y
+                x = self.output_netcdf.createVariable(
+                    'x', 'f4', self._netcdf_coords[1:])
+                y = self.output_netcdf.createVariable(
+                    'y', 'f4', self._netcdf_coords[1:])
+                x[:] = self.x
+                y[:] = self.y
+
+            else:
+                # new output format is 1d x and y
+                x = self.output_netcdf.createVariable(
+                    'x', 'f4', ('x'))
+                y = self.output_netcdf.createVariable(
+                    'y', 'f4', ('y'))
+                x[:] = self.xc
+                y[:] = self.yc
+
+            x.units = 'meter'
+            y.units = 'meter'
 
             # set up variables for output data grids
             def _create_grid_variable(varname, varunits,
@@ -696,7 +714,7 @@ class init_tools(abc.ABC):
         self._save_var_list['meta']['hb'] = ['hb', 'meters', 'f4', ()]
         self._save_var_list['meta']['cell_type'] = ['cell_type',
                                                     'type', 'i8',
-                                                    ('y', 'x')]
+                                                    self._netcdf_coords[1:]]
         # subsidence metadata
         if self._toggle_subsidence:
             self._save_var_list['meta']['start_subsidence'] = [
@@ -704,17 +722,20 @@ class init_tools(abc.ABC):
             ]
             self._save_var_list['meta']['sigma'] = [
                 'sigma', 'meters per timestep', 'f4',
-                ('y', 'x')
+                self._netcdf_coords[1:]
             ]
         # time-varying metadata
[…]
