### 2021-10-14 a0b52c9
change default and update tests, with checks on legacy version
