### 2021-10-14 05723a2
update and clean documentation
