+++ docs/source/examples/variable_velocity.rst
@@ -94,12 +94,12 @@ We define a model subclass to handle the changing boundary condition:
 and then simply run with:
 
 .. plot::
-    :context: close-figs
+    :context:
 
     # we create the model here, just to be sure it works (for good docs)
     with pyDeltaRCM.shared_tools._docs_temp_directory() as output_dir:
         mdl = ChangingVelocityModel(
-            end_time=end_time,
+            end_time=86400*100,
             out_dir=output_dir)
 
 .. code::
