+++ pyDeltaRCM/model.py
@@ -1285,7 +1285,7 @@ class DeltaModel(iteration_tools, sed_tools, water_tools,
         | dimensions  | `time`, `x`, `y`  | `total_time`, `length`, `width` |
         +-------------+-------------------+---------------------------------+
         | variables   | `time`, `x`, `y`  | `time`, `y`, `x`; x, y as 2D    |
-        +-------------+------------------------+----------------------------+
+        +-------------+-------------------+---------------------------------+
         | data        | `t-x-y` array     | `t-y-x` array                   |
         +-------------+-------------------+---------------------------------+
 
