+++ examples/subsidence_region-3.py
@@ -1,3 +1,0 @@
-fig, ax = plt.subplots()
-mdl.show_attribute('sigma', grid=False)
-plt.show()
