+++ guides/user_guide-1.py
@@ -1,2 +1,0 @@
-with pyDeltaRCM.shared_tools._docs_temp_directory() as output_dir:
-    mdl = pyDeltaRCM.DeltaModel(out_dir=output_dir)
