+++ tests/test_iteration_tools.py
@@ -715,3 +715,22 @@ class TestSaveGrids:
         assert _arr.shape[1] == _delta.qs.shape[0]
         assert _arr.shape[2] == _delta.qs.shape[1]
         assert ('meta' in ds.groups)  # if any grids, save meta too
+
+    def test_save_grids_exception(self, tmp_path):
+        p = utilities.yaml_from_dict(tmp_path, 'input.yaml', {'save_dt': 1})
+        _delta = DeltaModel(input_file=p)
+
+        # mock the log_info
+        _delta.log_info = mock.MagicMock()
+
+        # no netcdf will be created
+        exp_path_nc = os.path.join(
+            tmp_path / 'out_dir', 'pyDeltaRCM_output.nc')
+        assert os.path.isfile(exp_path_nc) is False
+
+        # trying to save a grid will raise an exception
+        with pytest.raises(Exception):
+            _delta.save_grids('sedflux', _delta.qs, _delta._save_iter)
+
+        # assert log was called when exception was raised
+        assert (_delta.log_info.call_count == 1)
