### 2021-11-29 9986499
add regex check for exception
